# doc/openapi.yaml
openapi: 3.0.3
info:
  title: CoZaDzban API
  description: |
    ## Overview
    CoZaDzban is a link-collecting application. Users submit a URL and a language
    tag; the server fetches the target page and automatically enriches the record
    with title, thumbnail, favicon, lead, and best-effort publication date before persisting it.

    ## Components
    1. **REST API** — CRUD endpoints for `Article` resources (this spec).
    2. **PostgreSQL database** — all data persisted via Spring Data JDBC.
    3. **Flyway migrations** — all schema changes are managed as versioned SQL
       migration scripts located in `src/main/resources/db/migration/`.
       No DDL is generated or executed outside of Flyway.
    4. **RSS feed** — `GET /rss` exposes articles as an RSS 2.0 feed;
       filterable per language with `?lang=<BCP-47>`.
    5. **Vaadin UI** — server-side lazy feed of articles (separate from
       this REST spec; rendered at the root path `/`).

    6. **Frontend authentication** — Vaadin UI starts OAuth Authorization Code + PKCE
       via backend redirect (`GET /auth/login`) and uses server-side session cookie.
    7. **Admin pages** — Vaadin UI at `/admin` for ADMIN users to manage allowlist
       (add/soft-delete/restore/update users) and at `/admin/article-content-cache`
       to inspect, refresh, edit, and purge preserved article content cache entries,
       using server-side service calls.
...
       imports remain non-interactive.
...
    ## Implementation notes
...

...
servers:
...
paths:
...
  /api/users:

    get:
      tags: [Security]
      summary: List allowlisted users
      operationId: listAppUsers
      x-required-role: ADMIN
      security:
        - googleOidc: []
      responses:
        '200':
          description: List of allowlisted users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Security]
      summary: Add user to allowlist
      operationId: addAppUser
      x-required-role: ADMIN
      security:
        - googleOidc: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppUserInput'
      responses:
        '201':
          description: User added to allowlist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/users/{id}/status:

    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: Allowed user identifier to update

    patch:
      tags: [Security]
      summary: Compatibility endpoint for updating user status
      operationId: updateAppUserStatus
      x-required-role: ADMIN
      security:
        - googleOidc: []
      description: Prefer PATCH /api/users/{id} with the status field for new clients.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppUserStatusPatch'
      responses:
        '200':
          description: User restored to ACTIVE
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppUser'
        '204':
          description: User marked DELETED
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
        description: Allowed user identifier to remove

    delete:
      tags: [Security]
      summary: Soft-delete user from allowlist
      operationId: removeAppUser
      x-required-role: ADMIN
      security:
        - googleOidc: []
      responses:
        '204':
          description: User marked DELETED and no longer allowed to authenticate or authorize writes
        '409':
          $ref: '#/components/responses/Conflict'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    patch:
      tags: [Security]
      summary: Update allowlisted user role/status
      operationId: updateAppUserRole
      x-required-role: ADMIN
      security:
        - googleOidc: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppUserRolePatch'
      responses:
        '200':
          description: Allowed user role/status updated (or unchanged if same values requested)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    googleOidc:
      type: openIdConnect
      openIdConnectUrl: https://accounts.google.com/.well-known/openid-configuration

    machineApiKey:
      type: apiKey
      in: header
      name: X-CoZaDzban-M2M-Key
      description: Machine credential used by the remote Facebook importer to write articles

    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Server-managed authenticated session cookie
...
  schemas:
...
    AppUserInput:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 254
          description: Email to add to the write-operation allowlist (RFC 5321 max 254 chars)
          example: admin@cozadzban.pl
        role:
          type: string
          enum: [USER, ADMIN]
          minLength: 1
          description: Authorization role assigned to this user
          example: USER
        status:
          type: string
          enum: [ACTIVE, DELETED]
          description: Optional initial lifecycle status; defaults to ACTIVE
          example: ACTIVE
    AppUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        email:
          type: string
          format: email
          maxLength: 254
          example: admin@cozadzban.pl
        role:
          type: string
          enum: [USER, ADMIN]
          example: USER
        status:
          type: string
          enum: [ACTIVE, DELETED]
          example: ACTIVE
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: "2026-05-02T10:00:00Z"
    AppUserRolePatch:
      type: object
      minProperties: 1
      description: Role/status update payload for allowlisted user
      properties:
        role:
          type: string
          enum: [USER, ADMIN]
          minLength: 1
          description: New role to assign
          example: ADMIN
        status:
          type: string
          enum: [ACTIVE, DELETED]
          description: New lifecycle status to assign
          example: ACTIVE
    AppUserStatusPatch:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ACTIVE, DELETED]
          description: New lifecycle status to assign
          example: ACTIVE
...
    ProblemDetail:
      type: object
      description: RFC 9457 / Spring ProblemDetail error response
      required: [type, title, status, detail]
      additionalProperties: true
      properties:
        type:
          type: string
          format: uri
          example: https://example.com/problems/validation-error
        title:
          type: string
          example: Bad Request
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "Field 'url' must be a valid URI"
        instance:
          type: string
          format: uri
          example: /api/articles
        conflictField:
          type: string
          description: Optional field name that caused a conflict
          example: url
        existingId:
          type: integer
          format: int64
          description: Optional existing resource identifier for conflict cases
          example: 42
        remainingAdmins:
          type: integer
          description: Optional count of admins that would remain after operation
          example: 0
  responses:
    BadRequest:
      description: Request payload is invalid or malformed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: Conflict with business rules
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
          examples:
            articleUrlConflict:
              summary: Duplicate article URL
              value:
                type: https://cozadzban.pl/problems/article-url-conflict
                title: Article URL Already Exists
                status: 409
                detail: "Article with this URL already exists (canonicalized: <url>)."
                instance: /api/articles
                conflictField: url
                existingId: 42
            allowlistEmailConflict:
              summary: Duplicate allowlist email
              value:
                type: https://cozadzban.pl/problems/allowlist-email-conflict
                title: Allowlisted Email Already Exists
                status: 409
                detail: User with this email is already allowlisted.
                instance: /api/users
                conflictField: email
                existingId: 7
            lastAdminRequired:
              summary: Attempt to remove final admin
              value:
                type: https://cozadzban.pl/problems/last-admin-required
                title: Last Admin Required
                status: 409
                detail: Operation would leave the system without an ADMIN user.
                instance: /api/users/1
                remainingAdmins: 0
            facebookImportBusy:
              summary: Facebook import already running
              value:
                type: https://cozadzban.pl/problems/facebook-import-busy
                title: Facebook Import Already Running
                status: 409
                detail: Facebook import is already running.
                instance: /api/admin/facebook-import/run
            facebookImportNotRunning:
              summary: No Facebook import to terminate
              value:
                type: https://cozadzban.pl/problems/facebook-import-not-running
                title: No Active Facebook Import
                status: 409
                detail: No Facebook import job is currently running.
                instance: /api/admin/facebook-import/terminate
    Unauthorized:
      description: Missing or invalid authentication credentials
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: Authenticated but not allowed to perform this operation
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: The requested resource does not exist
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...